+++ backend/routes/auth/login.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const User = require('../models/user');
+const User = require('../../models/user');
 
 router.get('/login', async (req, res) => {
   let jwt = 'sadsad';
